P0: split mapper_common into language-aware mappers (#25)
Move the language-specific dispatch tables and map_node_kind out of
mapper_common.py. Each mapper_<lang>.py now owns its own
_DECLARATION_TYPES / _STATEMENT_TYPES / _EXPRESSION_TYPES dicts and
its own map_node_kind, which it passes to the agnostic walker.

mapper_common.py now holds only the language-agnostic walker: span
construction, native-ref construction, child traversal, and the
generic to_uast helper that accepts a map_node_kind callable.

Pure refactor: per-language dispatch tables are verbatim copies of
the prior union table, so the UAST kind histograms for the
binarytrees benchmark are byte-identical before and after. Follow-up
issues will diverge the per-language tables to fix the silent
miss-matches called out in #25 (e.g. Python's binary_operator).

Closes #25
